requirements.txt
absl-py==0.6.1
astor==0.7.1
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
cloudpickle==0.7.0
cupy==5.3.0
cupy-cuda90==5.2.0
cycler==0.10.0
dask==1.1.1
decorator==4.3.2
fastrlock==0.4
Flask==1.1.1
gast==0.2.0
grpcio==1.17.1
h5py==2.9.0
idna==2.8
itsdangerous==1.1.0
Janome==0.3.7
Jinja2==2.10.3
kiwisolver==1.0.1
Markdown==3.0.1
MarkupSafe==1.1.1
matplotlib==3.0.2
mock==2.0.0
networkx==2.2
numpy==1.16.2
opencv-python==4.0.0.21
pandas==0.24.2
pbr==5.1.3
Pillow==6.2.2
protobuf==3.6.1
pydub==0.23.0
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.9
PyWavelets==1.0.1
requests==2.21.0
scikit-image==0.14.2
scikit-learn==0.20.2
scipy==1.2.0
selenium==3.141.0
six==1.12.0
sklearn==0.0
soupsieve==1.8
tensorboard==1.13.1
termcolor==1.1.0
toolz==0.9.0
torch==1.3.0
torchvision==0.4.1
urllib3==1.24.2
Werkzeug==0.16.0
